+++ packages/node/CHANGELOG.md
@@ -1,5 +1,15 @@
 # posthog-node
 
+## 5.26.0
+
+### Minor Changes
+
+- [#3144](https://github.com/PostHog/posthog-js/pull/3144) [`d43e90d`](https://github.com/PostHog/posthog-js/commit/d43e90df464b697a3aa0b46084d97a773a699abd) Thanks [@dustinbyrne](https://github.com/dustinbyrne)! - feat: getFeatureFlagResult, getAllFlags, getAllFlagsAndPayloads now have context-sensitive overrides which do not require `distinctId` as a parameter, instead reading it from the current context.
+  (2026-02-25)
+
+- [#3144](https://github.com/PostHog/posthog-js/pull/3144) [`d43e90d`](https://github.com/PostHog/posthog-js/commit/d43e90df464b697a3aa0b46084d97a773a699abd) Thanks [@dustinbyrne](https://github.com/dustinbyrne)! - feat: Add `enterContext()` to attach a context to the current async execution context
+  (2026-02-25)
+
 ## 5.25.0
 
 ### Minor Changes